+++ src/chain/definitions.rs
@@ -1,22 +1,21 @@
 //! Common objects for chain interaction system
 
-use jsonrpsee::ws_client::WsClient;
-use primitive_types::H256;
-use substrate_crypto_light::common::{AccountId32, AsBase58};
-use tokio::sync::oneshot;
-
 use crate::{
     chain::{
         rpc::{asset_balance_at_account, system_balance_at_account},
         tracker::ChainWatcher,
     },
     definitions::{
-        api_v2::{OrderInfo, Timestamp},
+        api_v2::{OrderInfo, RpcInfo, Timestamp},
         Balance,
     },
-    error::{ChainError, NotHex},
+    error::{ChainError, NotHexError},
     utils::unhex,
 };
+use jsonrpsee::ws_client::WsClient;
+use primitive_types::H256;
+use substrate_crypto_light::common::{AccountId32, AsBase58};
+use tokio::sync::oneshot;
 
 /// Abstraction to distinguish block hash from many other H256 things
 #[derive(Debug, Clone)]
@@ -25,14 +24,14 @@ pub struct BlockHash(pub primitive_types::H256);
 impl BlockHash {
     /// Convert block hash to RPC-friendly format
     pub fn to_string(&self) -> String {
-        format!("0x{}", hex::encode(&self.0))
+        format!("0x{}", const_hex::encode(&self.0))
     }
 
     /// Convert string returned by RPC to typesafe block
     ///
     /// TODO: integrate nicely with serde
     pub fn from_str(s: &str) -> Result<Self, crate::error::ChainError> {
-        let block_hash_raw = unhex(&s, NotHex::BlockHash)?;
+        let block_hash_raw = unhex(&s, NotHexError::BlockHash)?;
         Ok(BlockHash(H256(
             block_hash_raw
                 .try_into()
@@ -89,6 +88,7 @@ pub enum ChainRequest {
     WatchAccount(WatchAccount),
     Reap(WatchAccount),
     Shutdown(oneshot::Sender<()>),
+    GetConnectedRpcs(oneshot::Sender<Vec<RpcInfo>>),
 }
 
 #[derive(Debug)]