+++ src/state.rs
@@ -1,18 +1,15 @@
 use crate::{
     chain::ChainManager,
-    database::ConfigWoChains,
-    database::Database,
+    database::{ConfigWoChains, Database},
     definitions::api_v2::{
-        CurrencyProperties, OrderCreateResponse, OrderInfo, OrderQuery, OrderResponse, OrderStatus,
-        ServerInfo, ServerStatus,
+        CurrencyProperties, Health, OrderCreateResponse, OrderInfo, OrderQuery, OrderResponse,
+        OrderStatus, RpcInfo, ServerHealth, ServerInfo, ServerStatus,
     },
     error::{Error, OrderError},
     signer::Signer,
-    task_tracker::TaskTracker,
+    utils::task_tracker::TaskTracker,
 };
-
 use std::collections::HashMap;
-
 use substrate_crypto_light::common::{AccountId32, AsBase58};
 use tokio::sync::oneshot;
 use tokio_util::sync::CancellationToken;
@@ -82,7 +79,7 @@ impl State {
                         .add_invoice(order, order_details, state.recipient)
                         .await;
                 }
-                Ok("All saved orders restored".into())
+                Ok("All saved orders restored")
             });
 
             loop {
@@ -122,6 +119,15 @@ impl State {
                                 };
                                 res.send(server_status).map_err(|_| Error::Fatal)?;
                             }
+                            StateAccessRequest::ServerHealth(res) => {
+                                let connected_rpcs = state.chain_manager.get_connected_rpcs().await?;
+                                let server_health = ServerHealth {
+                                    server_info: state.server_info.clone(),
+                                    connected_rpcs: connected_rpcs.clone(),
+                                    status: Self::overall_health(&connected_rpcs),
+                                };
+                                res.send(server_health).map_err(|_| Error::Fatal)?;
+                            }
                             StateAccessRequest::OrderPaid(id) => {
                                 // Only perform actions if the record is saved in ledger
                                 match state.db.mark_paid(id.clone()).await {
@@ -158,14 +164,29 @@ impl State {
                 }
             }
 
-            Ok("State handler is shutting down".into())
+            Ok("State handler is shutting down")
         });
 
         Ok(Self { tx })
     }
 
+    fn overall_health(connected_rpcs: &Vec<RpcInfo>) -> Health {
+        if connected_rpcs.iter().all(|rpc| rpc.status == Health::Ok) {
+            Health::Ok
+        } else if connected_rpcs.iter().any(|rpc| rpc.status == Health::Ok) {
+            Health::Degraded
+        } else {
+            Health::Critical
+        }
+    }
+
     pub async fn connect_chain(&self, assets: HashMap<String, CurrencyProperties>) {
-        self.tx.send(StateAccessRequest::ConnectChain(assets)).await;
+        self.tx
+            .send(StateAccessRequest::ConnectChain(assets))
+            .await
+            .unwrap_or_else(|e| {
+                tracing::error!("Failed to send ConnectChain request: {}", e);
+            });
     }
 
     pub async fn order_status(&self, order: &str) -> Result<OrderResponse, Error> {
@@ -184,6 +205,15 @@ impl State {
         let (res, rx) = oneshot::channel();
         self.tx
             .send(StateAccessRequest::ServerStatus(res))
+            .await
+            .map_err(|_| Error::Fatal)?;
+        rx.await.map_err(|_| Error::Fatal)
+    }
+
+    pub async fn server_health(&self) -> Result<ServerHealth, Error> {
+        let (res, rx) = oneshot::channel();
+        self.tx
+            .send(StateAccessRequest::ServerHealth(res))
             .await
             .map_err(|_| Error::Fatal)?;
         rx.await.map_err(|_| Error::Fatal)
@@ -253,6 +283,7 @@ enum StateAccessRequest {
         res: oneshot::Sender<bool>,
     },
     ServerStatus(oneshot::Sender<ServerStatus>),
+    ServerHealth(oneshot::Sender<ServerHealth>),
     OrderPaid(String),
 }