+++ src/utils.rs
@@ -1,9 +1,13 @@
-use crate::error::{NotHex, UtilError};
+use crate::error::{NotHexError, UtilError};
 
-pub fn unhex(hex_data: &str, what_is_hex: NotHex) -> Result<Vec<u8>, UtilError> {
+pub mod logger;
+pub mod shutdown;
+pub mod task_tracker;
+
+pub fn unhex(hex_data: &str, what_is_hex: NotHexError) -> Result<Vec<u8>, UtilError> {
     if let Some(stripped) = hex_data.strip_prefix("0x") {
-        hex::decode(stripped).map_err(|_| UtilError::NotHex(what_is_hex))
+        const_hex::decode(stripped).map_err(|_| UtilError::NotHex(what_is_hex))
     } else {
-        hex::decode(hex_data).map_err(|_| UtilError::NotHex(what_is_hex))
+        const_hex::decode(hex_data).map_err(|_| UtilError::NotHex(what_is_hex))
     }
 }